requirements.txt
# Reconciliation Agent + Klaus Collections
# Merged requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
python-multipart==0.0.6

# Database (optional - not required for Railway deployment)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# API Integrations - Reconciliation
plaid-python==16.0.0
hubspot-api-client==12.0.0
anthropic==0.7.7

# API Integrations - Klaus
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Background Tasks & Scheduling
celery==5.3.4
redis==5.0.1
apscheduler>=3.10.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Fuzzy Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytz>=2023.3

# Notifications (optional)
twilio>=8.0.0

# CORS
fastapi-cors==0.0.6

# Optional: For advanced features
# openpyxl>=3.1.0  # For Excel export
# sentry-sdk>=1.0.0  # For error tracking